requirements.txt
Django==5.2.3
djangorestframework
django-cors-headers
django-countries
django-filter
djangorestframework-simplejwt
Pillow
gunicorn
coverage
python-decouple
psycopg2-binary
whitenoise